Disable rules for making target build of libbcc
These rules add stuff to $(ALL_MODULES.libbcc.REQUIRED), which affects
the host build as well (this seems to be a bug/misfeature in the build
system).  Since we're only building the host variant here, and don't
want to bother building prerequisites for the target variants, just
eliminate the rules for the target build.